lib/uksched: Rework waitqueues to use thread data
This is a ground-up reimplementation of wait queues to use wakeup
entries (tickets) embedded within the thread data structure instead of
allocated on the execution stack of a thread as before.
This ensures a thread is waiting on at most one queue at a time, and
enables detection when terminating a waiting thread.

The public wait queue API remains unchanged.

Checkpatch-Ignore: MACRO_ARG_REUSE
Checkpatch-Ignore: SPACING
Signed-off-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1651